+++ packages/generator/generators/app/templates/package.json
@@ -21,40 +21,24 @@
   ],
   "author": "Talend Frontend <frontend@talend.com>",
   "license": "Apache-2.0",
-  "devDependencies": {
-    "@talend/scripts": "^0.18.0"
-  },
   "dependencies": {
-    "@talend/bootstrap-theme": "3.0.0",
-    "@talend/icons": "3.0.0",
-    "@talend/react-cmf": "3.0.0",
-    "@talend/react-components": "3.0.0",
-    "@talend/react-containers": "3.0.0",
-    "bootstrap-sass": "3.4.1",
-    "bson-objectid": "^1.1.5",
-    "classnames": "^2.2.5",
-    "enzyme": "^3.7.0",
+    "@talend/bootstrap-theme": "3.4.0",
+    "@talend/icons": "3.4.0",
+    "@talend/locales-tui": "3.1.0",
+    "@talend/react-cmf": "3.4.0",
+    "@talend/react-components": "3.4.0",
+    "@talend/react-containers": "3.4.0",
+    "focus-visible": "^4.1.5",
     "i18next": "^9.0.0",
-    "immutable": "^3.8.1",
-    "invariant": "^2.2.2",
-    "lodash": "^4.17.4",
     "prop-types": "^15.5.10",
     "react": "^16.8.5",
-    "react-bootstrap": "0.32.4",
-    "react-css-transition": "^0.7.4",
     "react-dom": "^16.8.5",
     "react-i18next": "^7.6.1",
-    "react-immutable-proptypes": "^2.1.0",
     "react-redux": "^5.0.7",
-    "react-router": "^3.2.0",
-    "react-router-redux": "^4.0.8",
-    "react-transition-group": "^2.3.1",
-    "redux": "^3.7.2",
-    "redux-batched-actions": "^0.2.0",
-    "redux-logger": "^3.0.6",
-    "redux-saga": "^0.15.4",
-    "redux-thunk": "^2.2.0",
-    "reselect": "^2.5.4",
-    "uuid": "^3.0.1"
+    "redux": "^3.7.2"
+  },
+  "devDependencies": {
+    "@talend/scripts": "^0.18.0",
+    "enzyme": "^3.7.0"
   }
 }
