+++ packages/generator/generators/app/templates/src/app/components/App/App.component.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { I18nextProvider } from 'react-i18next';
 import IconsProvider from '@talend/react-components/lib/IconsProvider';
 
+import i18n from '../../i18n';
 import Home from '../Home';
 
 export default function App() {
@@ -11,10 +13,12 @@ export default function App() {
 	 * so that all icons are available in each view
 	 */
 	return (
-		<React.Fragment>
-			<IconsProvider />
-			<Home />
-		</React.Fragment>
+		<I18nextProvider i18n={i18n}>
+			<React.Fragment>
+				<IconsProvider />
+				<Home />
+			</React.Fragment>
+		</I18nextProvider>
 	);
 }
