+++ packages/generator/generators/app/templates/src/app/components/DatasetList/index.js
@@ -1,3 +1,20 @@
+import { connect } from 'react-redux';
 import DatasetList from './DatasetList.component';
+import DatasetService from '../../services/dataset';
 
-export default DatasetList;
+function mapStateToProps(state) {
+	return {
+		datasets: DatasetService.selectors.getDatasets(state),
+		isFetching: DatasetService.selectors.getIsFetching(state),
+	};
+}
+
+const mapDispatchToProps = {
+	fetchDatasets: DatasetService.actionCreators.fetchDatasets,
+	removeDataset: DatasetService.actionCreators.removeDataset,
+};
+
+export default connect(
+	mapStateToProps,
+	mapDispatchToProps,
+)(DatasetList);
