+++ packages/generator/generators/app/templates/src/app/index.css
@@ -1,0 +1,8 @@
+/*
+  @see https://github.com/WICG/focus-visible
+  This will hide the focus indicator if the element receives focus via the mouse,
+  but it will still show up on keyboard focus.
+*/
+.js-focus-visible :focus:not(.focus-visible) {
+	outline: none;
+}
